requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.4.0
PyYAML==6.0.2
rich==13.9.2
scikit_learn==1.1.1
statsmodels==0.13.2
torch==1.11.0
torchvision==0.20.0
tqdm==4.66.5
argparse==1.4.0
gdown==4.6.1